Collect the data and group them by user id

+++ src/main/java/edu/snu/cms/bdcs/assignment/data/Rate.java
@@ -7,7 +7,7 @@ public interface Rate {
   /**
    * @return The rate value
    */
-  public long getRate();
+  public byte getRate();
 
   /**
    * @return The user id who rated this rate value
